Create network metadata hpp
